app/requirements.txt
PyOpenGL
pygame
requests
flask
